pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sleep-disorder"
version = "0.1.0"
description = "Package to process raw data and train a classifier."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mlflow==3.4.0",
    "numpy==2.3.3",
    "pandas==2.3.3",
    "PyYAML==6.0.3",
    "requests==2.32.5",
    "scikit-learn==1.7.2"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true
